Unable to produce the JAR file appropriately.
Main no longer extends Application, used a wrapper class labeled as
'MainWrapper'.
