+++ docs/README.md
@@ -2,7 +2,7 @@
 
 This is Duke, a chat bot to help you track your tasks.
 
-## Features 
+## Features
 
 - Adding a Todo task: `todo`
 - Adding an Event task: `event`
@@ -17,15 +17,14 @@ This is Duke, a chat bot to help you track your tasks.
 
 ### `todo DESCRIPTION` - Adds a todo task with the specified description.
 
-Example of usage: 
+Example of usage:
 
 `todo TODO EXAMPLE`
 
 Expected outcome:
 
-```
-<img src="todo.png" width="420" />
-```
+![image](https://user-images.githubusercontent.com/97376457/189971908-5fe12440-c4c9-4027-a5ec-bc6216d13a30.png)
+
 ### `event DESCRIPTION /at DATE` - Adds an event task with the specified description and date.
 - Ensure date is in format <YYYY-MM-DD>
 
@@ -35,9 +34,7 @@ Example of usage:
 
 Expected outcome:
 
-```
-<img src="event.png" width="420" />
-```
+![image](https://user-images.githubusercontent.com/97376457/189972627-567efaf2-2378-4f46-aef8-50f1a874867e.png)
 
 ### `deadline DESCRIPTION /by DATE` - Adds a deadline task with the specified description and date.
 - Ensure date is in format <YYYY-MM-DD>
@@ -48,9 +45,7 @@ Example of usage:
 
 Expected outcome:
 
-```
-<img src="deadline.png" width="420" />
-```
+![image](https://user-images.githubusercontent.com/97376457/189972673-22841325-9d06-4e6b-ba64-48d730bb101c.png)
 
 ### `list` - Lists all tasks currently saved in the list.
 
@@ -60,9 +55,7 @@ Example of usage:
 
 Expected outcome:
 
-```
-<img src="list.png" width="420" />
-```
+![image](https://user-images.githubusercontent.com/97376457/189972738-018e190b-c060-4e48-8a23-17df2dd878c4.png)
 
 ### `find KEYWORD` - Returns a list of tasks containing the keyword specified.
 
@@ -72,9 +65,7 @@ Example of usage:
 
 Expected outcome:
 
-```
-<img src="find.png" width="420" />
-```
+![image](https://user-images.githubusercontent.com/97376457/189972765-78fb576e-732b-4d9f-a3fe-7911d0caf477.png)
 
 ### `delete INDEX` - Deletes the task in the list at that given index.
 
@@ -84,9 +75,7 @@ Example of usage:
 
 Expected outcome:
 
-```
-<img src="delete.png" width="420" />
-```
+![image](https://user-images.githubusercontent.com/97376457/189972818-00eee5f5-3f1f-4419-93aa-7381242142fd.png)
 
 ### `mark INDEX` - Marks the task as done in the list at that given index.
 
@@ -96,9 +85,7 @@ Example of usage:
 
 Expected outcome:
 
-```
-<img src="mark.png" width="420" />
-```
+![image](https://user-images.githubusercontent.com/97376457/189972845-c3b7b24a-cef3-4309-b8f3-a8fd2dbeb7b3.png)
 
 ### `unmark INDEX` - Marks the task as not done in the list at that given index.
 
@@ -108,9 +95,7 @@ Example of usage:
 
 Expected outcome:
 
-```
-<img src="unmark.png" width="420" />
-```
+![image](https://user-images.githubusercontent.com/97376457/189972880-2b1e5d72-a48b-4946-a80e-90d94a1f849a.png)
 
 ### `tag INDEX #HASHTAG` - Tags a task with a #hashtag at that given index.
 
@@ -120,9 +105,7 @@ Example of usage:
 
 Expected outcome:
 
-```
-<img src="tag.png" width="420" />
-```
+![image](https://user-images.githubusercontent.com/97376457/189972898-ca871fc5-9eb4-4e33-bfe8-8f496c6d783b.png)
 
 ### `bye` - Saves your current list and then closes the program.
 
@@ -135,4 +118,3 @@ Expected outcome:
 ```
 The program terminates by itself.
 ```
-
