+++ src/main/java/duke/Main.java
@@ -1,32 +1,11 @@
 package duke;
 
-import java.io.IOException;
-
-import duke.gui.MainWindow;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
 /**
  * A GUI for Duke using FXML.
  */
-public class Main extends Application {
+public class Main{
 
-    private Duke duke = new Duke();
-
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    public static void main(String[] args) {
+        MainWrapper.main(args);
     }
 }
