+++ src/main/java/duke/gui/Launcher.java
@@ -1,6 +1,6 @@
 package duke.gui;
 
-import duke.Main;
+import duke.MainWrapper;
 import javafx.application.Application;
 
 /**
@@ -8,6 +8,6 @@ import javafx.application.Application;
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Main.class, args);
+        Application.launch(MainWrapper.class, args);
     }
 }
